implement function on next and previous buttons
the buttons now go to the next and previous entries in the shapes list
and displays them in the painter.

if the list is empty or we are at the last index the list is appended
if we are at any point inside of the list the shape is replaced with the
new details

Note: the functionality is a litte bit awkward when we try to replace
the last
Note2: for improved functionality we can update the values in the
dropdowns and the spinboxes, I dont think this is required in the scope
of the assignment.